# contract/openapi.yaml
...
components:
  schemas:
...
    FieldValidationError:
      description: "검증 실패 필드 한 건 (422 응답의 errors[] 원소)"
      properties:
        field:
          description: "실패한 요청 필드 이름"
          type: "string"
        message:
          description: "한국어 오류 메시지"
          type: "string"
      required:
      - "field"
      - "message"
      type: "object"
...
    Problem:
      description: "RFC 9457 오류 응답 + Pickle 확장(code, errors). 모든 비정상 응답은 이 형태로 반환됩\
        니다."
      properties:
        code:
          description: "기계 판독용 안정 오류 코드 (클라이언트 분기 기준)"
          type: "string"
        detail:
          description: "이 발생 건에 대한 설명"
          type: "string"
        errors:
          description: "검증 실패 상세 (422에서만 존재)"
          items:
            $ref: "#/components/schemas/FieldValidationError"
          type: "array"
        instance:
          description: "발생 경로"
          type: "string"
        status:
          description: "HTTP 상태 코드"
          format: "int32"
          type: "integer"
        title:
          description: "사람이 읽는 짧은 제목"
          type: "string"
        type:
          description: "문제 유형 URI (기본 about:blank)"
          type: "string"
      required:
      - "title"
      - "status"
      - "code"
      type: "object"
...
    SshKeyAlgorithm:
      enum:
      - "ED25519"
      - "RSA"
      type: "string"
    SshKeyCreateRequest:
      properties:
        name:
          maxLength: 100
          minLength: 1
          type: "string"
        publicKey:
          maxLength: 8192
          minLength: 0
          type: "string"
      required:
      - "name"
      - "publicKey"
      type: "object"
    SshKeyGenerateRequest:
      properties:
        name:
          maxLength: 100
          minLength: 1
          type: "string"
      required:
      - "name"
      type: "object"
    SshKeyPrivateKeyResponse:
      properties:
        fileName:
          type: "string"
        privateKey:
          type: "string"
      required:
      - "fileName"
      - "privateKey"
      type: "object"
    SshKeyView:
      properties:
        algorithm:
          $ref: "#/components/schemas/SshKeyAlgorithm"
        createdAt:
          format: "date-time"
          type: "string"
        fingerprint:
          type: "string"
        id:
          format: "uuid"
          type: "string"
        lastUsedAt:
          format: "date-time"
          type:
          - "string"
          - "null"
        name:
          type: "string"
        privateKeyStored:
          type: "boolean"
        publicKey:
          type: "string"
      required:
      - "algorithm"
      - "createdAt"
      - "fingerprint"
      - "id"
      - "name"
      - "privateKeyStored"
      - "publicKey"
      type: "object"
...
    TerminalTicketResponse:
      properties:
        expiresAt:
          format: "date-time"
          type: "string"
        sessionId:
          type: "string"
        subprotocol:
          type: "string"
        ticket:
          type: "string"
        wsPath:
          type: "string"
      required:
      - "expiresAt"
      - "sessionId"
      - "subprotocol"
      - "ticket"
      - "wsPath"
      type: "object"
...
  securitySchemes:
    bearerAuth:
      bearerFormat: "JWT"
      description: "액세스 토큰 (JWT HS256, 15분 만료). 클레임: sub(계정 공개 식별자 UUID), role, token_version.\
        \ 비밀번호 변경·계정 비활성화 시 token_version이 올라가 기존 토큰이 즉시 무효화됩니다. 리프레시 토큰은 보안 스킴이 아\
        니라 __Host-pickle_refresh httpOnly 쿠키로만 오갑니다."
      scheme: "bearer"
      type: "http"
info:
  description: "부산대학교 클라우드 플랫폼 Pickle의 REST API. 인증은 JWT Bearer, 오류 응답은 RFC 9457 problem+json(Problem\
    \ 스키마)을 따릅니다."
  title: "Pickle API"
  version: "0.38.0"
openapi: "3.1.0"
paths:
...
  /me/ssh-keys:
    get:
      operationId: "listKeys"
      responses:
        "200":
          content:
            '*/*':
              schema:
                items:
                  $ref: "#/components/schemas/SshKeyView"
                type: "array"
          description: "OK"
        default:
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: "오류 — 상태 코드와 무관하게 Problem 형태"
      tags:
      - "me-ssh-key-controller"
    post:
      operationId: "registerKey"
      parameters:
      - description: "재인증(sudo-mode) 토큰 — POST /auth/reverify가 발급 (10분 유효, 다회용). 없\
          거나 만료·무효면 403 REAUTH_REQUIRED."
        in: "header"
        name: "X-Reauth-Token"
        required: false
        schema:
          type: "string"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SshKeyCreateRequest"
        required: true
      responses:
        "201":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/SshKeyView"
          description: "Created"
        "403":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: "재인증 필요 — 유효한 X-Reauth-Token 없음 (`REAUTH_REQUIRED`)"
        default:
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: "오류 — 상태 코드와 무관하게 Problem 형태"
      tags:
      - "me-ssh-key-controller"
  /me/ssh-keys/generate:
    post:
      operationId: "generateKey"
      parameters:
      - description: "재인증(sudo-mode) 토큰 — POST /auth/reverify가 발급 (10분 유효, 다회용). 없\
          거나 만료·무효면 403 REAUTH_REQUIRED."
        in: "header"
        name: "X-Reauth-Token"
        required: false
        schema:
          type: "string"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SshKeyGenerateRequest"
        required: true
      responses:
        "201":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/SshKeyView"
          description: "Created"
        "403":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: "재인증 필요 — 유효한 X-Reauth-Token 없음 (`REAUTH_REQUIRED`)"
        default:
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: "오류 — 상태 코드와 무관하게 Problem 형태"
      tags:
      - "me-ssh-key-controller"
  /me/ssh-keys/{keyId}:
    delete:
      operationId: "deleteKey"
      parameters:
      - in: "path"
        name: "keyId"
        required: true
        schema:
          format: "uuid"
          type: "string"
      - description: "재인증(sudo-mode) 토큰 — POST /auth/reverify가 발급 (10분 유효, 다회용). 없\
          거나 만료·무효면 403 REAUTH_REQUIRED."
        in: "header"
        name: "X-Reauth-Token"
        required: false
        schema:
          type: "string"
      responses:
        "204":
          description: "No Content"
        "403":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: "재인증 필요 — 유효한 X-Reauth-Token 없음 (`REAUTH_REQUIRED`)"
        default:
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: "오류 — 상태 코드와 무관하게 Problem 형태"
      tags:
      - "me-ssh-key-controller"
  /me/ssh-keys/{keyId}/private-key:
    get:
      operationId: "downloadPrivateKey"
      parameters:
      - in: "path"
        name: "keyId"
        required: true
        schema:
          format: "uuid"
          type: "string"
      - description: "재인증(sudo-mode) 토큰 — POST /auth/reverify가 발급 (10분 유효, 다회용). 없\
          거나 만료·무효면 403 REAUTH_REQUIRED."
        in: "header"
        name: "X-Reauth-Token"
        required: false
        schema:
          type: "string"
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/SshKeyPrivateKeyResponse"
          description: "OK"
        "403":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: "재인증 필요 — 유효한 X-Reauth-Token 없음 (`REAUTH_REQUIRED`)"
        default:
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: "오류 — 상태 코드와 무관하게 Problem 형태"
      tags:
      - "me-ssh-key-controller"
...
  /vms/{vmId}/terminal-sessions:
    post:
      operationId: "createTerminalSession"
      parameters:
      - in: "path"
        name: "vmId"
        required: true
        schema:
          format: "uuid"
          type: "string"
      responses:
        "201":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/TerminalTicketResponse"
          description: "Created"
        default:
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: "오류 — 상태 코드와 무관하게 Problem 형태"
      tags:
      - "terminal-controller"
...
security:
- bearerAuth: []
servers:
- url: "/api/v1"
...